+++ Source/Gui/Dialogs/HotkeyEditor.cpp
@@ -76,8 +76,8 @@ LRESULT CHotkeyEditor::OnInitDialog(UINT uMsg, WPARAM wParam, LPARAM lParam, BOO
 
 LRESULT CHotkeyEditor::OnClickedOK(WORD wNotifyCode, WORD wID, HWND hWndCtl, BOOL& bHandled)
 {	
-	m_data.localKey = ConvertHotkeyToPair(localHotkeyCtrl.GetWinHotkey());
-	m_data.globalKey = ConvertHotkeyToPair(globalHotkeyCtrl.GetWinHotkey());
+	m_data.localKey = ConvertHotkeyToPair(static_cast<WORD>(localHotkeyCtrl.GetWinHotkey()));
+	m_data.globalKey =ConvertHotkeyToPair( static_cast<WORD>(globalHotkeyCtrl.GetWinHotkey()));
 	DoDataExchange(TRUE);
 	EndDialog(wID);
 	return 0;
